requirements.txt
alembic==0.9.10
amqp==2.5.0
apache-airflow==1.10.3
appnope==0.1.0
asn1crypto==0.24.0
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.7.0
backcall==0.1.0
billiard==3.6.0.0
boto3==1.7.84
botocore==1.10.84
celery==4.3.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
Click==7.0
cloudant==0.5.10
colorama==0.4.1
-e git+ssh://git@code.scripbox.io:12022/scripbox/mars.git@5277d2e9f6650d2933f38f4121397b8d1a321d32#egg=common&subdirectory=dags/common
configparser==3.5.3
croniter==0.3.30
cryptography==2.7
Cython==0.29.10
decorator==4.4.0
defusedxml==0.6.0
dill==0.2.9
docutils==0.14
entrypoints==0.3
flake8==3.7.8
Flask==1.0.3
Flask-Admin==1.5.3
Flask-AppBuilder==1.12.3
Flask-Babel==0.12.2
Flask-Caching==1.3.3
Flask-Login==0.4.1
Flask-OAuthlib==0.9.5
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.0
flask-swagger==0.2.13
Flask-WTF==0.14.2
funcsigs==1.0.0
future==0.16.0
future-fstrings==1.1.0
gitdb2==2.0.5
GitPython==2.1.11
gunicorn==19.9.0
-e git+ssh://git@code.scripbox.io:12022/scripbox/mars.git@5277d2e9f6650d2933f38f4121397b8d1a321d32#egg=hiveql&subdirectory=hiveql
hmsclient==0.1.1
idna==2.8
importlib-metadata==0.18
ipython==7.5.0
ipython-genutils==0.2.0
iso8601==0.1.12
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
json-merge-patch==0.2
kombu==4.6.1
lockfile==0.12.2
lxml==4.3.4
Mako==1.0.12
Markdown==2.6.11
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
mysqlclient==1.3.14
numpy==1.16.4
oauthlib==2.1.0
ordereddict==1.1
packaging==19.0
pandas==0.24.2
parso==0.4.0
pendulum==1.4.4
pexpect==4.7.0
pickleshare==0.7.5
pinotdb==0.1.1
pluggy==0.12.0
prompt-toolkit==2.0.9
protobuf==3.8.0
psutil==5.6.3
psycopg2==2.7.7
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pydruid==0.5.5
pyflakes==2.1.1
Pygments==2.4.2
PyHive==0.6.1
pymongo==3.8.0
pyparsing==2.4.0
pytest==4.6.3
pytest-helpers-namespace==2019.1.8
python-daemon==2.1.2
python-dateutil==2.8.0
python-editor==1.0.4
python3-openid==3.1.0
pytz==2019.1
pytzdata==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-futures==0.9.4
requests-oauthlib==1.1.0
s3transfer==0.1.13
setproctitle==1.1.10
six==1.12.0
smmap2==2.0.5
snakebite==2.11.0
SQLAlchemy==1.2.19
tabulate==0.8.3
tenacity==4.12.0
text-unidecode==1.2
thrift==0.11.0
traitlets==4.3.2
tzlocal==1.5.1
unicodecsv==0.14.1
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
zipp==0.5.1
zope.deprecation==4.4.0
